Fractional Knapsack
The fractional knapsack problem is a classic problem in combinatorial optimization. If a set of items are given, each with a weight and a value, the goal is to select a subset of the items that maximises the value while keeping the total weight below or equal to a given limit.